[ALPS07553785] Serialize access to GLB_DB_REQ to ring CSG doorbell
ARM Ticket : 00479758
patch : 4494

This commit fixes the race between the ring of CSG doorbell to send the
acknowledgment for the CS_FAULT event and the ring of CSG doorbell for
sending the CSG start/resume request. Due to the race the update made to
GLB_DB_REQ for the start/resume request was lost and so the request timed
out.

The acknowledgment for the CS_FAULT event is sent in interrupt handler
whereas CSG requests are sent in the process context where Scheduler
mutex is held. To serialize the access to GLB_DB_REQ between interrupt
handler and process context, this commit uses the interrupt_lock.

MTK-Commit-Id: f0fc08091ce1bd1808d18cb86e0186598c8bd9c7

Signed-off-by: Lian Lee <lian.lee@mediatek.com>
CR-Id: ALPS07553785
Feature: ARM GPU DDK driver
Change-Id: Idbc6ac0b57e8b7effdcceccfa8708ed4db1f4fd2