[ALPS08072316] [c5433]Fix for ringbuffer alloc failure after fatal error
[Detail]
ARM Ticket: 03425035
ARM Patch: https://code.ipdelivery.arm.com/c/GPU/mali-ddk/+/5433

When a fatal error occurs for a GPU queue/group, Kbase sends the error
notification to the Userspace. As a part of error recovery, Userspace
deletes the existing queues & groups and creates new ones.
The fatal error is handled in the bottom half and a reference is taken
on the GPU queue before enqueuing the work item to handle the error.
SAME_VA zone is used for the GPU queue ringbuffer and the memory for
ringbuffer remains pinned until the refcount of GPU queue drops to 0.
If Userspace tries to free the ringbuffer memory whilst it is pinned,
then only the CPU virtual range is freed but the GPU virtual range
remains allocated, due to which CPU VA space can go out of sync with
the GPU VA space.

It could happen that Userspace, after receiving the error notification
for one of the queues, start the error recovery actions and thereby
delete the queues whilst the bottom half to handle the fatal error for
the other queues is still pending/running. This can cause the creation
of new GPU queues to fail if the same GPU VA value cant be used for
the new queue as the CPU VA value chosen for it. Due to failure, error
recovery wont happen and Application would crash.

As a fix, this commit updates the handling of CS_QUEUE_TERMINATE ioctl
to block the Application thread till the bottom half for handling the
fatal error or Tiler OoM event is complete.

A Base defect test has also been added to verify the creation of new
queues do not fail after deleting the existing queues on fatal error.

MTK-Commit-Id: 37fb43edc9ca08319917e0bb9de6add01fc8ea08

CR-Id: ALPS08072316
Feature: ARM GPU DDK driver
Change-Id: I1d5d5e74695452944d055c822c40deadbb4ba970