[ALPS08279981] r44p1: align csg suspend timeout with r38p1
[Detail]
Because of the MTK freqency information wont be input mali driver for
timeout value calculation so the mali driver will use default 100Mhz
to have the calculation.
Then, it will end up with the original CSG suspend timeout to be 31s.
So, adjust the value to align with r38p1 1.5s.

MTK-Commit-Id: 67094b7140ef586ad13f6f501ecb81cfef6762bc

CR-Id: ALPS08279981
Feature: ARM GPU DDK driver
Change-Id: Ibe2ca6e218595353264b0102dfce51588546cddb