[ALPS08818877] GPU: GPUCORE-41385 for CVE-2024-0671
[Description]
Previously, when creating a GPU mapping, the driver would only consider
the alignment of the GPU VA and the tag attached to the corresponding
physical page to decide whether to create a single MMU ATE at level 2
for a 2 MB region or a set of MMU ATEs at level 3 of 4 kB each.

The problem was that a small memory alias of a large memory allocation
could satisfy both conditions if its GPU VA happens to be aligned to
a large memory page. This would lead to the creation of a 2 MB mapping
even if the user only requested a smaller region, e.g. 4 kB. Moreover,
this would lead to problem during the teardown, leaving the MMU page
table in an inconsistent state and potentially keeping the GPU mapping
in existence, allowing use-after-free of a physical page after the
original allocation is released.

The driver now adds a third condition to the creation of a MMU ATE
at level 2: the total number of MMU page table entries to map must be
512, i.e. the user has requested a whole 2 MB region intentionally.

[Patch]
(c6487/1) Fix small memory aliases of large memory allocations

[ARM Ticket]
03464100

[Test]
CTS, Stress test

MTK-Commit-Id: 9be51c9414d43261325e1c3831d242f1eb17f73d

Signed-off-by: Daniel Tao <daniel.tao@mediatek.com>
CR-Id: ALPS08818877
Feature: ARM GPU DDK driver
Change-Id: If2499e89b904803229fc26ded1e5096834a4fdd8
(cherry picked from commit 4e45acbcddeecd41288ce4cf5ca224078894e64b)